+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 3.0.0-alpha.2 (2021-02-09)
+
+[CHANGED] Changed Broadcast method into BroadcastAsync
+[NEW] Added async/await
+[CHANGED] Made channel authentication more secure
+
 ## 3.0.0-alpha.1 (2021-02-09)
 
 [NEW] Work on base code of package, compatible with the next release of PushRadar